+++ src/app/components/core/notifications/add-notification-wizard/add-notification-wizard.component.ts
@@ -22,6 +22,7 @@ import { ValidateFormService } from '../../../../services/validate-form.service'
 export class AddNotificationWizardComponent implements OnInit, OnDestroy {
   @ViewChild('desc', { static: false }) description: ElementRef;
   @ViewChild('name', { static: false }) name: ElementRef;
+  @ViewChild('retriggerTime', { static: false }) retriggerTime: ElementRef;
 
   public notificationRulePlugins = [];
   public notificationDeliveryPlugins = [];
@@ -52,7 +53,8 @@ export class AddNotificationWizardComponent implements OnInit, OnDestroy {
     name: new FormControl(),
     description: new FormControl(),
     rule: new FormControl(),
-    delivery: new FormControl()
+    delivery: new FormControl(),
+    retriggerTime: new FormControl()
   });
 
   @ViewChild('ruleConfigView', { static: false }) ruleViewConfigItemComponent: ViewConfigItemComponent;
@@ -81,7 +83,8 @@ export class AddNotificationWizardComponent implements OnInit, OnDestroy {
       name: ['', Validators.required],
       description: ['', Validators.required],
       rule: ['', Validators.required],
-      delivery: ['', Validators.required]
+      delivery: ['', Validators.required],
+      retriggerTime: ['60', Validators.required]
     });
     this.subscription = this.sharedService.showLogs.subscribe(showPackageLogs => {
       if (showPackageLogs.isSubscribed) {
@@ -292,6 +295,10 @@ export class AddNotificationWizardComponent implements OnInit, OnDestroy {
           this.payload.notification_type = this.notificationType;
         }
         this.payload.enabled = this.isNotificationEnabled;
+        this.payload.retrigger_time = this.retriggerTime.nativeElement.value;
+        if (this.payload.retrigger_time < 1) {
+          return;
+        }
         this.addNotificationInstance(this.payload);
         break;
       default:
@@ -351,7 +358,7 @@ export class AddNotificationWizardComponent implements OnInit, OnDestroy {
     this.isRulePlugin = true;
     this.isDeliveryPlugin = true;
     const plugin = (selectedPlugin.slice(3).trim()).replace(/'/g, '');
-    if (pluginType === 'notificationRule') {
+    if (pluginType === 'rule') {
       this.selectedRulePluginDescription = this.notificationRulePlugins
         .find(p => p.config.plugin.default === plugin).config.plugin.description;
     } else {