+++ src/app/components/core/scheduler/list-schedules/list-schedules.component.ts
@@ -7,6 +7,7 @@ import { AlertDialogComponent } from '../../../common/alert-dialog/alert-dialog.
 import { UpdateScheduleComponent } from '../update-schedule/update-schedule.component';
 
 enum weekDays {
+  None = null,
   Monday = 1,
   Tuesday = 2,
   Wednesday = 3,
@@ -49,7 +50,6 @@ export class ListSchedulesComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-
     this.getScheduleType();
     this.getSchedulesProcesses();
 
@@ -105,8 +105,9 @@ export class ListSchedulesComponent implements OnInit {
         (data: any) => {
           /** request completed */
           this.ngProgress.done();
+          // north_c - North C tasks, north_C - Northbound C Services
           data.schedules.forEach(sch => {
-            if (!['south_c', 'north_c', 'south', 'north', 'notification_c'].includes(sch.processName)) {
+            if (!['south_c', 'north_c', 'north_C', 'south', 'north', 'notification_c'].includes(sch.processName)) {
               this.scheduleData.push(sch);
             }
           });
@@ -120,6 +121,7 @@ export class ListSchedulesComponent implements OnInit {
               element.repeat = repeatTimeObj.time;
             }
             element.time = Utils.secondsToDhms(element.time).time;
+            element.dayName = weekDays[element.day];
           });
           this.scheduleData = sortBy(this.scheduleData, function (obj) {
             return !obj.enabled + obj.name.toLowerCase();